+++ serde-generate/runtime/golang/src/lcs/lcs_test.go
@@ -713,7 +713,7 @@ func TestSerializeDeserializeOptionTag(t *testing.T) {
 func TestGetBufferOffset(t *testing.T) {
 	s := new(lcs.Serializer)
 	s.SerializeU64(0)
-	assert.Equal(t, s.GetBufferOffset(), 8)
+	assert.Equal(t, s.GetBufferOffset(), uint64(8))
 	d := lcs.NewDeserializer([]byte{0, 0, 0, 0, 0, 0, 0, 0})
-	assert.Equal(t, d.GetBufferOffset(), 8)
+	assert.Equal(t, d.GetBufferOffset(), uint64(8))
 }
